+++ generator-core/src/algos/map_builder/mod.rs
@@ -1,187 +1,25 @@
-use crate::{constants::REGION_SPLIT_FACTOR, types::*};
+use crate::{
+    MapStyle,
+    types::{Map, Rect},
+};
 
 use anyhow::Result;
 use rayon::prelude::*;
 use tracing::event;
 
 mod add_doors;
+mod bisect_rooms;
 mod bsp;
+mod builder_config;
 mod connect_regions;
 mod gen_rooms;
+mod merge_regions;
 mod merge_rooms;
 mod reconnect_rooms;
 mod room_decorator;
 
-pub(crate) struct MapBuilderConfig {
-    pub bsp_config: bsp::BinarySpacePartitioningConfig,
-    // Should we merge the regions after generating their rooms?
-    pub merge_regions: bool,
-    // The probability of randomly merging two rooms into one.
-    pub random_room_merge_prob: f64,
-    // Probability of having a group reconnect to two groups instead of one
-    pub group_loop_connection_chance: f64,
-    // Probability of opening a connection between rooms that will
-    // cause a navigation loop in the map.
-    pub loop_connection_chance: f64,
-    pub repeat_small_room_merge_prob: f64,
-    pub bisect_room_prob: f64,
-}
-
-impl Default for MapBuilderConfig {
-    fn default() -> Self {
-        MapBuilderConfig {
-            bsp_config: bsp::BinarySpacePartitioningConfig::default(),
-            merge_regions: true,
-            random_room_merge_prob: 0.05,
-            group_loop_connection_chance: 0.17,
-            loop_connection_chance: 0.2,
-            repeat_small_room_merge_prob: 0.2,
-            bisect_room_prob: 0.1,
-        }
-    }
-}
-
-impl MapBuilderConfig {
-    pub fn from_style(style: MapStyle) -> Self {
-        let mut base = Self::default();
-
-        match style {
-            MapStyle::CastlevaniaSOTN => {
-                base.bsp_config.horizontal_region_prob = 0.75;
-                base.bsp_config.big_rect_area_cutoff = 14;
-                base.bsp_config.big_rect_survival_prob = 0.09;
-                base.bsp_config.horizontal_split_prob = 0.85;
-                base.bsp_config.height_factor_cutoff = 2.9;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.33;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.8;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.9;
-
-                base.random_room_merge_prob = 0.03;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.22;
-                base.repeat_small_room_merge_prob = 0.51;
-                base.bisect_room_prob = 0.17;
-            }
-            MapStyle::CastlevaniaAOS => {
-                base.bsp_config.horizontal_region_prob = 0.0;
-                base.bsp_config.big_rect_area_cutoff = 11;
-                base.bsp_config.big_rect_survival_prob = 0.12;
-                base.bsp_config.horizontal_split_prob = 0.82;
-                base.bsp_config.height_factor_cutoff = 2.4;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.51;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.77;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.85;
-
-                base.random_room_merge_prob = 0.01;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.24;
-                base.repeat_small_room_merge_prob = 0.45;
-                base.bisect_room_prob = 0.15;
-            }
-            MapStyle::CastlevaniaCOTM => {
-                base.bsp_config.horizontal_region_prob = 0.1;
-                base.bsp_config.big_rect_area_cutoff = 12;
-                base.bsp_config.big_rect_survival_prob = 0.15;
-                base.bsp_config.horizontal_split_prob = 0.82;
-                base.bsp_config.height_factor_cutoff = 1.4;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.95;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.95;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.60;
-
-                base.random_room_merge_prob = 0.15;
-                base.group_loop_connection_chance = 0.10;
-                base.loop_connection_chance = 0.14;
-                base.repeat_small_room_merge_prob = 0.85;
-                base.bisect_room_prob = 0.29;
-            }
-            MapStyle::CastlevaniaHOD => {
-                base.bsp_config.horizontal_region_prob = 0.75;
-                base.bsp_config.big_rect_area_cutoff = 8;
-                base.bsp_config.big_rect_survival_prob = 0.09;
-                base.bsp_config.horizontal_split_prob = 0.85;
-                base.bsp_config.height_factor_cutoff = 1.9;
-                base.bsp_config.width_factor_cutoff = 1.6;
-                base.bsp_config.rect_survival_prob = 0.70;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.8;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.9;
-
-                base.random_room_merge_prob = 0.03;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.22;
-                base.repeat_small_room_merge_prob = 0.81;
-                base.bisect_room_prob = 0.17;
-            }
-            MapStyle::MetroidZM => {
-                base.bsp_config.region_split_factor =
-                    (REGION_SPLIT_FACTOR / 4) + (REGION_SPLIT_FACTOR / 2);
-                base.bsp_config.horizontal_region_prob = 0.75;
-                base.bsp_config.big_rect_area_cutoff = 14;
-                base.bsp_config.big_rect_survival_prob = 0.09;
-                base.bsp_config.horizontal_split_prob = 0.85;
-                base.bsp_config.height_factor_cutoff = 2.9;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.33;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.6;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.7;
-
-                base.merge_regions = false;
-
-                base.random_room_merge_prob = 0.03;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.22;
-                base.repeat_small_room_merge_prob = 0.51;
-                base.bisect_room_prob = 0.17;
-            }
-            MapStyle::MetroidFS => {
-                base.bsp_config.region_split_factor =
-                    (REGION_SPLIT_FACTOR / 4) + (REGION_SPLIT_FACTOR / 2);
-                base.bsp_config.horizontal_region_prob = 0.75;
-                base.bsp_config.big_rect_area_cutoff = 14;
-                base.bsp_config.big_rect_survival_prob = 0.09;
-                base.bsp_config.horizontal_split_prob = 0.85;
-                base.bsp_config.height_factor_cutoff = 2.9;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.33;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.8;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.9;
-
-                base.merge_regions = false;
-
-                base.random_room_merge_prob = 0.03;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.22;
-                base.repeat_small_room_merge_prob = 0.51;
-                base.bisect_room_prob = 0.17;
-            }
-            MapStyle::MetroidSP => {
-                base.bsp_config.region_split_factor =
-                    (REGION_SPLIT_FACTOR / 4) + (REGION_SPLIT_FACTOR / 2);
-                base.bsp_config.horizontal_region_prob = 0.75;
-                base.bsp_config.big_rect_area_cutoff = 14;
-                base.bsp_config.big_rect_survival_prob = 0.09;
-                base.bsp_config.horizontal_split_prob = 0.85;
-                base.bsp_config.height_factor_cutoff = 2.9;
-                base.bsp_config.width_factor_cutoff = 2.6;
-                base.bsp_config.rect_survival_prob = 0.33;
-                base.bsp_config.trim_highly_connected_rect_prob = 0.8;
-                base.bsp_config.trim_fully_connected_rect_prob = 0.9;
-
-                base.merge_regions = false;
-
-                base.random_room_merge_prob = 0.03;
-                base.group_loop_connection_chance = 0.19;
-                base.loop_connection_chance = 0.22;
-                base.repeat_small_room_merge_prob = 0.51;
-                base.bisect_room_prob = 0.17;
-            }
-        }
-
-        base
-    }
-}
+use builder_config::BinarySpacePartitioningConfig;
+pub(crate) use builder_config::MapBuilderConfig;
 
 pub(crate) struct MapBuilder {
     pub cols: u32,
@@ -200,27 +38,26 @@ impl MapBuilder {
     }
 
     pub fn build(&self, config: &MapBuilderConfig, style: MapStyle) -> Vec<Map> {
-        let build_start = std::time::SystemTime::now();
+        let build_start = std::time::Instant::now();
 
-        let rect_regions = bsp::BinarySpacePartitioning::generate_and_trim_partitions(
+        let rect_groups = bsp::BinarySpacePartitioning::generate_and_trim_partitions(
             self.cols,
             self.rows,
             config.bsp_config,
         );
 
-        let region_idx_offsets = Self::generate_region_offsets(
-            rect_regions.iter().map(|(_, rects, _, _)| rects),
-            rect_regions.len(),
-        );
+        // We generate index offsets so that when we merge the groups later,
+        // we don't have to worry about colliding room IDs.
+        let group_idx_offsets = Self::generate_group_offsets(&rect_groups);
 
-        let bsp_time = std::time::SystemTime::now();
+        let rect_groups_time = std::time::Instant::now();
         event!(
             tracing::Level::DEBUG,
-            "Generated BSP regions in {:?}ms",
-            bsp_time.duration_since(build_start).unwrap().as_millis()
+            "Generated rectangle groups in {:.2}ms",
+            rect_groups_time.duration_since(build_start).as_millis()
         );
 
-        let map_regions = rect_regions.into_par_iter().zip(region_idx_offsets).map(
+        let map_regions = rect_groups.into_par_iter().zip(group_idx_offsets).map(
             |((origin_rect, region_rects, removed_rects, neighbours), idx_offset)| {
                 let mut map_region =
                     Self::generate_map_region(origin_rect, region_rects, removed_rects, neighbours);
@@ -244,46 +81,60 @@ impl MapBuilder {
                     config.repeat_small_room_merge_prob / 2.0,
                 );
 
+                // Finally we bisect long horizontal rooms randomly
+                //MapBuilder::bisect_long_horizontal_rooms(&mut map_region, config.bisect_room_prob);
+
                 map_region
             },
         );
 
-        let map_regions_time = std::time::SystemTime::now();
+        let map_regions_time = std::time::Instant::now();
         event!(
             tracing::Level::DEBUG,
-            "Generated map regions in {:?}ms",
+            "Generated map regions in {:.2}ms",
             map_regions_time
-                .duration_since(bsp_time)
-                .unwrap()
+                .duration_since(rect_groups_time)
                 .as_millis()
         );
 
         let generated_maps = if config.merge_regions {
-            let origin_rect = Rect {
-                origin: Cell::new(0, 0),
-                width: self.cols,
-                height: self.rows,
-            };
+            let origin_rect = Rect::new(0, 0, self.cols, self.rows);
 
-            let mut map_region = map_regions.reduce(
-                || MapRegion::new(origin_rect),
-                |mut acc, region| {
-                    acc.rooms.extend(region.rooms);
-                    acc.removed_rooms.extend(region.removed_rooms);
-                    acc.neighbours.extend(region.neighbours);
+            let mut map_region = Self::merge_regions(origin_rect, map_regions, &config);
 
-                    acc
-                },
+            let merge_time = std::time::Instant::now();
+            event!(
+                tracing::Level::DEBUG,
+                "Merged map regions in {:.2}ms",
+                merge_time.duration_since(map_regions_time).as_millis()
             );
 
+            // We connect the rooms of the newly merged region together
             Self::reconnect_room_groups(&mut map_region, config);
 
-            //let doors = Self::add_doors_to_rooms(&map_region, config);
-            let doors = vec![];
+            let merge_reconnect_time = std::time::Instant::now();
+            event!(
+                tracing::Level::DEBUG,
+                "Reconnected rooms in {:.2}ms",
+                merge_reconnect_time.duration_since(merge_time).as_millis()
+            );
 
-            let room_decorator = room_decorator::RoomDecoratorFactory::decorator_for(style);
-            room_decorator::RoomDecorator::decorate(
-                room_decorator.as_ref(),
+            // At this point we have no more use for the removed rooms
+            // so we can clear them.
+            map_region.clear_removed_rooms();
+
+            let clear_removed_time = std::time::Instant::now();
+            event!(
+                tracing::Level::DEBUG,
+                "Cleared removed rooms in {:.2}ms",
+                clear_removed_time
+                    .duration_since(merge_reconnect_time)
+                    .as_millis()
+            );
+
+            let doors = Self::add_doors_to_rooms(&map_region, config);
+
+            room_decorator::RoomDecoratorFactory::decorator_for(style).decorate(
                 &mut map_region,
                 &doors,
                 config,
@@ -299,11 +150,13 @@ impl MapBuilder {
         } else {
             let mut maps = map_regions
                 .map(|mut map_region| {
+                    // At this point we have no more used for the removed rooms
+                    // so we can clear them.
+                    map_region.clear_removed_rooms();
+
                     let doors = Self::add_doors_to_rooms(&map_region, config);
 
-                    let room_decorator = room_decorator::RoomDecoratorFactory::decorator_for(style);
-                    room_decorator::RoomDecorator::decorate(
-                        room_decorator.as_ref(),
+                    room_decorator::RoomDecoratorFactory::decorator_for(style).decorate(
                         &mut map_region,
                         &doors,
                         config,
@@ -322,13 +175,12 @@ impl MapBuilder {
             maps
         };
 
-        let generated_maps_time = std::time::SystemTime::now();
+        let generated_maps_time = std::time::Instant::now();
         event!(
             tracing::Level::DEBUG,
-            "Generated maps in {:?}ms",
+            "Added doors and modifiers in {:.2}ms",
             generated_maps_time
                 .duration_since(map_regions_time)
-                .unwrap()
                 .as_millis()
         );
 
@@ -339,14 +191,12 @@ impl MapBuilder {
         let built_doors = generated_maps
             .iter()
             .fold(0_usize, |acc, map| acc + map.doors.len());
-
-        let build_end = std::time::SystemTime::now();
         event!(
             tracing::Level::DEBUG,
-            "Built map with {} rooms and {} doors in {:?}ms",
+            "Built {} map/s with {} rooms and {} doors",
+            generated_maps.len(),
             built_rooms,
-            built_doors,
-            build_end.duration_since(build_start).unwrap().as_millis()
+            built_doors
         );
 
         generated_maps
